+++ climpred/tests/test_bootstrap.py
@@ -1,6 +1,7 @@
 import time
 
 import dask
+import numpy as np
 import pytest
 import xarray as xr
 from xarray.testing import assert_allclose
@@ -9,13 +10,13 @@ from climpred.bootstrap import (
     _bootstrap_by_stacking,
     bootstrap_hindcast,
     bootstrap_perfect_model,
-    bootstrap_uninit_pm_ensemble_from_control,
     bootstrap_uninit_pm_ensemble_from_control_cftime,
     my_quantile,
 )
 from climpred.comparisons import HINDCAST_COMPARISONS, PM_COMPARISONS
-from climpred.constants import VALID_ALIGNMENTS
+from climpred.constants import CONCAT_KWARGS, VALID_ALIGNMENTS
 from climpred.exceptions import KeywordError
+from climpred.utils import _transpose_and_rechunk_to
 
 BOOTSTRAP = 2
 
@@ -133,6 +134,65 @@ def test_bootstrap_hindcast_resample_dim(
     )
 
 
+def bootstrap_uninit_pm_ensemble_from_control(init_pm, control):
+    """
+    Create a pseudo-ensemble from control run. Deprecated in favor of
+    `bootstrap_uninit_pm_ensemble_from_control_cftime`.
+
+    Note:
+        Needed for block bootstrapping confidence intervals of a metric in perfect
+        model framework. Takes randomly segments of length of ensemble dataset from
+        control and rearranges them into ensemble and member dimensions.
+
+    Args:
+        init_pm (xarray object): ensemble simulation.
+        control (xarray object): control simulation.
+
+    Returns:
+        uninit (xarray object): pseudo-ensemble generated from control run.
+    """
+    nens = init_pm.init.size
+    nmember = init_pm.member.size
+    length = init_pm.lead.size
+    c_start = 0
+    c_end = control['time'].size
+    lead_time = init_pm['lead']
+
+    def set_coords(uninit, init, dim):
+        uninit[dim] = init[dim].values
+        return uninit
+
+    def isel_years(control, year_s, length):
+        new = control.isel(time=slice(year_s, year_s + length))
+        new = new.rename({'time': 'lead'})
+        new['lead'] = lead_time
+        return new
+
+    def create_pseudo_members(control):
+        startlist = np.random.randint(c_start, c_end - length - 1, nmember)
+        uninit_ens = xr.concat(
+            (isel_years(control, start, length) for start in startlist),
+            dim='member',
+            **CONCAT_KWARGS,
+        )
+        return uninit_ens
+
+    uninit = xr.concat(
+        (
+            set_coords(create_pseudo_members(control), init_pm, 'member')
+            for _ in range(nens)
+        ),
+        dim='init',
+        **CONCAT_KWARGS,
+    )
+    # chunk to same dims
+    return (
+        _transpose_and_rechunk_to(uninit, init_pm)
+        if dask.is_dask_collection(uninit)
+        else uninit
+    )
+
+
 def test_bootstrap_uninit_pm_ensemble_from_control_cftime_annual_identical(
     PM_ds_initialized_1d_ym_cftime, PM_ds_control_1d_ym_cftime
 ):
@@ -141,6 +201,7 @@ def test_bootstrap_uninit_pm_ensemble_from_control_cftime_annual_identical(
     cftime_res = bootstrap_uninit_pm_ensemble_from_control_cftime(
         PM_ds_initialized_1d_ym_cftime, PM_ds_control_1d_ym_cftime
     )
+
     noncftime_res = bootstrap_uninit_pm_ensemble_from_control(
         PM_ds_initialized_1d_ym_cftime, PM_ds_control_1d_ym_cftime
     )
